Move ConnectionRef into the two pools
